requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
urllib3==2.0.6
